JavaScript: implement Typed Arrays, namely ArrayBuffer, DataView, Float32Array, Float64Array, Int16Array, Int32Array, Int8Array, Uint16Array, Uint32Array, Uint8Array and Uint8ClampedArray.
Issue 3564695

git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@7517 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/javascript/host/arrays/Uint32Array.java
@@ -64,7 +64,7 @@ public class Uint32Array extends ArrayBufferViewBase {
     protected Long fromArray(final byte[] array, final int offset) {
         final ByteBuffer buff = ByteBuffer.wrap(array);
         buff.order(ByteOrder.LITTLE_ENDIAN);
-        return ((long) buff.getInt(offset)) & 0xFFFFFFFFL;
+        return buff.getInt(offset) & 0xFFFFFFFFL;
     }
 
     /**
